patchman/octree: when inquiring for octant data ghost and internal octants shall be considered separately
A new structure called 'OctantInfo' is introduced, this allow
to specify if the octant is an internal octant or a ghost. The
functions that inquire octant data now use the new 'OctantInfo'
data structure.

Moreover, we need to keep two distinct cell to octant maps: one
for the internal octants and one for the ghost octants.