patchman/octree: fix deletion of cells
The previous version of the code was wrong, because
when removing an element from the cell-to-octant
or the octant-to-cell maps, it was using the value
as the argument of the erase function (we should
use the key instead).

Instead of just fix the issue, the code has been
rework to make it simpler.