patchman/piercedVector: add a function to extract the flat id of an element
A flat id is the id associated to a numbering scheme that starts
from the element in the first position of the container and is
incremented by one for each element in the container. The first
element will have a flat id equal to 0, the last element will
have a flat id equal to (nElements - 1).