patchman/piercedVector: add functions to "reclaim" elements from the piercedVector
These functions get an element from a position marked as empty and
assign to it the specified id. Except for setting the id, the element
is not modified. Therefore it will still contain	the data of the
element that was previously occupying the position or it will be
empty if there was no empty position and a new element has been created.