test: simplify handling of parallel tests
Now there is just one list for both serial and parallel tests.
If the name of a test contains the word "parallel", the test
will be considered a parallel test.

By default, parallel test will be executed using the number of
processors defined by the variable 'MPIEXEC_MAX_NUMPROCS'. If a
different number of processors is required, this number can be
specified when adding the test to the lists of tests.  The number
of processors to be used for the test has to be appended to the
test name separated by a colon (:) character.