patch: implement 'moveInternal2Ghost' and 'moveGhost2Internal' using the swap function
When moving an internal to the ghosts, the internal is swapped with
the last internal and then first ghost and last internal id markers
are updated. In a similar way, when moving a ghosts to the internals,
the ghost is swapped with the first ghosts and then the markers are
updated.