containers/piercedVector: rework handling of the holes
All the holes are stored in a single vector. The first part of this
vector contains the "regular" holes, whereas the last part contains
the "pending" holes. The space reserved to the pending holes is fixed.
When this space if full, the 'holes_flush' function will be called
and all pending holes will be converted to regular holes.