containers/piercedVector: try to optimize the update of empty elements when flushing the holes
The list of pending holes is sorted, in this way we can iterate
from the last pending hole in the container to the first one.
We start updating the id of the last pending hole, updating also
the id of the contiguous holes before that one. Then we advance
to the next hole, skipping the positions that have already been
updated.