PABLO: fix search of neighbours for ghosts that are processor boundaries
We always need to search in the internals, the only exception is
for faces of internal octants that are processor boundaries.

If we also want neighbours that are ghosts, we always need to search
in the ghosts, the only exception is for faces of internal octants
that are not processor boundaries.