PABLO: always search face neighbours in the internal octants
A face is considered on a processors boundary when at least one
of its neighbours is owned by a different rank. This means that,
if a face is on a processors boundary, it's not guaranteed that
all its neighbours will be ghosts. Therefore, we always need to
search face neighbours among the internal octants.