PABLO: fix evaluation of the bounding box
On two-dimensional trees, the z coordinate of the bounding box
should be set equal to the z coordinate of the origin.

If the tree has no octants the bounding box is empty.

When scanning the nodes of the tree for evaluating the bounding
box, it is possible to get node coordinates directly from the
'PabloUniform' class.