RBF: changed base class structure
Added an abstract base class retaining all the feature of the old RBF, except
for the type of nodes used in the RBF parameterization. The new RBF class
derived from it, explicitly uses as nodes 3D points.
Old functionality of RBF are guaranteed. You can now derive a new RBF class using
a different type of node. For example an unstructured mesh can be used as RBF node,
the only thing to be provided are a suitable calculation of distances.