VTK: error or warnings are written to log file if the size of data written or read is suspicous
The advancement of the stream is monitored during read/write of field data.

For writing the correct advancement can be known a-priori only binary formats and an error message is written in the log and code will stop in debug mode.
For ascii data a the error message is written and code is stopped if NO advancement is registrated.

For reading only warnings are thrown since the user may not be interested in reading the field.