containers: allocate the internal buffer of 'BinaryStream' in chunks
The 'setCapacity' function will guarantee that the capacity of the
buffer will always contain an integer number of chunks and that the
number of chunks will fit in an 'int' type of variable. To obtain
this, when the requested capacity is greater than the maximum integer
value, the buffer is allocated in chunks of 2^n bytes, where n is
choosen to obtain a number of chunks that fits in an 'int' type of
variable.