communications: communicate data in chucks
This commits allows the DataCommunicator to handle data transfers
bigger than 2Gb. The number of elements communicated by MPI has to
fit in an 'int' type of variable. To overcome this limit, the data
is communicated in data chunks. The BinaryBuffer will take care of
choosing the proper chunk size that makes the number of chunks fit
in an 'int' type of variable.