containers: rework 'BinaryStream' size allocation
There is now a distinction between the size of the stream and the
capacity of the stream. The size is the length of the data contained
in the stream (or the length of the data the stream will contain, if
not all the data has been inserted yet). The capacity is the storage
space allocated for the stream.

Capacity is not necessarily equal to the stream size; it can be equal
or greater. The extra space can not be used for storing data, but
will be used in the next commit to guarantee a minimum internal storage
size.