cmake: allow to specify a list of required/optional modules when invoking find_package
A list of required BITPIT modules may be specified when invoking the
find_package command after the COMPONENTS option (or after the REQUIRED
option if present). Additional optional components may be listed after
OPTIONAL_COMPONENTS. For each of the requested modules, a boolean variable
named BITPIT_<MODULE_NAME>_FOUND will be set telling if the corresponding
module is enabled in the corresponding BITPIT installation. If a required
module is not found a fatal error is generated and the configure step
stops executing.