containers/piercedVector: catch exception in swap function of PiercedVector
The 'swap' function of the PiercedVector is declared as 'noexcept',
so it cannot throw any exceptions. The function that sets a dynamic
kernel may throw an exception if the kernel is already set or if
we are trying to set a null kernel. Here, neither of the two cases can
happen, because the kernel has been previously cleared and the kernel
we are trying to set is not null.