patchkernel: optimize deletion of vertices
Deleting the last vertex requires some additional work. If the ids
of vertices to be deleted contain the last vertex, that vertex
is deleted after deleting all other vertices. In this way we make
sure to delete the last vertex just once (after deleting the last
vertex, another vertex becomes the last one and that vertex
may be on the deletion list as well, and on and so forth).