IO: VTK/VTKUnstructured seekg/tellg fix for Win-MinGW compiling
The commit addressed the problem of stream repositioning with
seekg/tellg when the stream is opened in full text mode(ascii). Because
of different escaping chars w.r.t to Unix, seekg/tellg on ascii file
under Windows do not work as presumed in bitpit::VTK code.
The trick is to open always the file in binary mode: the last works both
in win and unix/linux.

See notes for developers left as code comment, for usage examples.