cmake: add functions to define deprecated/experimental features
Experimental feature are disabled by default, whereas deprecated
features are enabled by default.

Cmake variables related to experimental and deprecated functions are
advanced variables and they will only be visible if the variables
BITPIT_SHOW_EXPERIMENTAL or BITPIT_SHOW_DEPRECATED are set to ON.

For each deprecated/experimental feature a corresponding preprocessor
macro will be defined. Given the feature FEATURE in the module MODULE,
the corresponding preprocessor macro will be called BITPIT_MODULE_FEATURE.
The value of the macro will be set to 1 if the feature is enable or
it will be set to 0 if the feature is disabled.