patchkernel: optimize transfer of vertices into the patch
It is not possible to directly transfer a vertex into the storage.
First a new vertex should be created and then that vertex can be
replaced with the source that needs to be transferred into the patch.
In order to improve performances, the created vertex can be just a
dummy vertex, no need to copy its properties from the source.