patchkernel: optimize transfer of cells into the patch
It is not possible to directly transfer a cell into the storage.
First a new cell should be created and then that cell can be
replaced with the source that needs to be transferred into the patch.
In order to improve performances, the created cell can be just a
dummy cell, no need to copy its properties from the source.