patchkernel: optimize transfer of interfaces into the patch
It is not possible to directly transfer a interface into the storage.
First a new interface should be created and then that interface can be
replaced with the source that needs to be transferred into the patch.
In order to improve performances, the created interface can be just a
dummy interface, no need to copy its properties from the source.