patchkernel: fix ordering of polyhedra vertices
The list of polyhedra vertices should contain the vertices sorted in
the same order they first appear in the face stream. This guarantees
that vertices will be sorted in the same order regardless of the id
of the vertices. In this way given an element, the list of its
vertices generated by different processes will iterate the vertices
in the same order.