levleset: allow to chose between a dense or a sparse storage type
Levelset can use two types of storages: sparse or dense. When sparse storage
is used, objects will only allocate space for storing information of cells
that are inside the narrow band. Instead, when dense storage is used, objects
will still evaluate information only for cells inside the narrow band, but
they will allocate space for all cells of the mesh. If the narrow band covers
a large portion of the domain, it will be more efficient to allocate space
for all the cells of the mesh, rather than keep track of the cells inside
the narrow band and store space only for those cells.