levelset: fix search radius for Cartesian meshes
Search radius should be equal to the maximum between the narrow band
size and the diameter of the circumcircle. This guarantees that, when
the narrow band size is equal or less than zero, the levelset will be
evaluated on the cells that intersect the surface and on all their
neighbours.