general: MSVC manual exporting Global Static data
CMAKE_WINDOWS_EXPORT_ALL_SYMBOLS fails in exporting global static data (static or extern defined in hpp). Explicit __declspec is done to cure this flaw.
A private compiler definition BITPIT_DLLGLOBALDATA_EXPORT is defined to mark this data as dllexport during the library/subtargets building and to declare them as dllimport when they are used  in exes or libraries which links to bitpit (when their hpp is included).