containers: fix PiercedStorage SFINAE usage
The usage of T in declaration of method's arguments triggers an error
since MSVC does not resolve T=value_t=<pod specified by the class> in
instantiation. In this context, T is used just as a dummy to trigger
SFINAE, so it's safe to use directly original class value_t as method
argument.