LA: allow to set the block size of a linear system
Rather than working with individual elements in the matrix, it is possible
to employ blocks of elements. The size of the blocks can be defined during
assembly. When a size different that one is provided, the matrix will store
elements by fixed-sized dense nb × nb blocks, where nb is the size of the
blocks. Blocking may be advantageous when solving PDE-based simulations
that leads to matrices with a naturally blocked structure (with a block size
equal to the number of degrees of freedom per cell).