containers/piercedVector: allow to pass a constant kernel when setting dynamic synchronization
The rationale for this change is that when a storage is registered for
synchronization, it won't alter the kernel in a way that affects the
usage of the kernel. What happens is that a new PiercedSyncSlave will
be added to the list of objects that should be keep in sync with the
kernel.