containers/piercedKernel: fix the update of ids associated with empty positions
The id of an empty element contains the distance, measured in number of elements, between
the current element and the next element the iterator should jump to (the distance is
negative). If the next position is non-empty or a regular hole, the value of the id will
be set in such a way to point to the next non-empty element (this will make the iterator
as fast as possible). If the next position is a pending hole, we need to set the id to
-1, otherwise we may break the iterator. For example,

      ID    1  2  3 -1  4 -1 -1  5
     POS    0  1  2  3  4  5  6  7

when setting the id associated with position 4, it should be set to -1 and not to -3,
otherwise if the position 5 will be filled, the iterator will not able to reach it (it
will reach position 4 and will wrongly skip to position 7).