levelset: micro-optimize sign propagation
If the sign of the cell has already been set, we check if it matches
the seed sign and then we skip the cell, otherwise we set the sign of
the cell and we process its neighoburs. Once the sign of a cell is
set, we can remove it form the seed list (there is no need to start
again the porpagation from that cell, beause it will not reach any
new portions of the domain).